refactor(dotfiles_py): move around project structure
