+++ dotfiles_py/data/scripts/colors.sh
