+++ dotfiles_py/data/scripts/config.sh
