+++ dotfiles_py/data/scripts/nvim_time.sh
