+++ dotfiles_py/data/scripts/select_nvim.sh
