+++ dotfiles_py/targets/lint.py
@@ -5,7 +5,7 @@ import tempfile
 
 from git import Repo
 
-from ..utils import REPO_PATH, SCRIPTS_PATH, run_shell, yaml_read
+from ..utils import DOTPY_PATH, REPO_PATH, run_shell, yaml_read
 from .ansible_collections import ANSIBLE_COLLECTIONS_PATH, ansible_collections
 from .roles_graph import roles_graph
 
@@ -156,13 +156,13 @@ def lint_code(*, ansible: bool, python: bool, secrets: bool, generic: bool) -> N
             extra_env={"ANSIBLE_COLLECTIONS_PATH": ANSIBLE_COLLECTIONS_PATH},
         )
         run_shell(
-            ["ansible-lint", SCRIPTS_PATH / "playbook_dotfiles_container.yaml"],
+            ["ansible-lint", REPO_PATH / "playbook_dotfiles_container.yaml"],
             extra_env={"ANSIBLE_COLLECTIONS_PATH": ANSIBLE_COLLECTIONS_PATH},
         )
     if python:
-        run_shell(["python3", "-m", "mypy", REPO_PATH / "scripts"])
+        run_shell(["python3", "-m", "mypy", DOTPY_PATH])
         # Specifying job number for pylint somehow leads to false-positive errors
-        run_shell(["python3", "-m", "pylint", REPO_PATH / "scripts"])
+        run_shell(["python3", "-m", "pylint", DOTPY_PATH])
         run_shell(["python3", "-m", "yamllint", "--strict", REPO_PATH / ".github"])
 
     if generic:
