+++ main.sh
@@ -87,4 +87,4 @@ if ! command -v dot &>/dev/null; then
   fi
 fi
 
-uv run "$root/__main__.py" "${@:---help}"
+uv run dotfiles "${@:---help}"
