--- roles/terminals/files/base46.lua
@@ -1,5 +1,0 @@
-return {
-  override_theme = function(x, _)
-    return x
-  end,
-}
